+++ baseutils/src/test/java/moe/xing/baseutils/utils/FileUtilsTest.java
@@ -1,0 +1,18 @@
+package moe.xing.baseutils.utils;
+
+import org.junit.Test;
+
+/**
+ * Created by Qi Xingchen on 2017/11/22.
+ * <p>
+ * FileUtilsTest {@link FileUtils}
+ */
+public class FileUtilsTest {
+    @Test
+    public void getFileNameFromUrl() throws Exception {
+
+//        URL url = new URL("http://www.example.com/some/path/to/a/file.xml#233?foo=bar#test");
+//        assertEquals("file.xml", FileUtils.getFileNameFromUrl(url.toString()));
+    }
+
+}